_archive/requirements_python.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
google-genai>=0.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
